+++ src/components/SidebarShareModeModal/index.jsx
@@ -2,9 +2,8 @@ import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
-import deleteCookie from "../../../utils/deleteCookie";
 import { SERVER_ADDRESS } from "../../../utils/env";
-import getCookie from "../../../utils/getCookie";
+import { getCookie, deleteCookie } from "../../../utils/manageCookie";
 import COLORS from "../../constants/COLORS";
 import { setShareKey } from "../../redux/reducers/shareKey";
 import { setUrlAddress } from "../../redux/reducers/urlAddress";
