--- utils/deleteCookie.js
@@ -1,3 +1,0 @@
-export default (name) => {
-  document.cookie = name + "=; expires=Thu, 01 Jan 1999 00:00:10 GMT;";
-};
