--- utils/getCookie.js
@@ -1,5 +1,0 @@
-export default (name) => {
-  const value = document.cookie.match("(^|;) ?" + name + "=([^;]*)(;|$)");
-
-  return value ? value[2] : null;
-};
